backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
blinker==1.9.0
boto3==1.40.50
botocore==1.40.50
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.1.8
colorama==0.4.6
constantly==23.10.4
crashtest==0.4.1
cryptography==46.0.2
cssselect2==0.8.0
daphne==4.2.1
distlib==0.4.0
distro==1.9.0
Django==5.2.3
django-axes==8.0.0
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-flex-fields==1.0.2
drf-spectacular==0.28.0
dulwich==0.22.8
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==37.4.2
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.59.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huey==2.5.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
lxml==6.0.0
MarkupSafe==3.0.3
more-itertools==10.7.0
msgpack==1.1.1
numpy==2.3.3
oauthlib==3.3.1
openai==1.97.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pbs-installer==2025.7.23
pdf2image==1.17.0
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMuPDF==1.26.3
pyOpenSSL==25.3.0
pyparsing==3.2.3
pypdf==5.9.0
pyphen==0.17.2
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.4.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.5.9.12
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
virtualenv==20.32.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.4
whitenoise==6.9.0
xhtml2pdf==0.2.17
zope.interface==8.0.1
zopfli==0.2.3.post1
zstandard==0.23.0
